Enable Go workspace
Introduce a new Go workspace that references all the current submodules.
To preserve the current behavior, point all of the current
FORBIDDEN_DEPENCENCY virtual dependencies to the same path.

Add scripts/update_go_workspace.sh that generates the Go workspace file
(go.work) based on the submodules found in the project (excluding the
tools) and creates the go.work.sum file after running go mod download.
Added this script to the fix Makefile target. Even though, the number of
modules in the etcd repository is small, by adding an automated way of
updating the go.work modules future proofs the project in case there are
new modules or removal of them.

Remove the workaround to forbid dependencies through the
FORBIDDEN_DEPENDENCY replace. This can be done through a linter rather
than at the go.mod level. Using the go.mod approach also brings issues
with the workspace, as all the dependencies (including replaces) need to
have a consistent version and point to the same place.

Co-authored-by: Tim Hockin <thockin@google.com>
Signed-off-by: Ivan Valdes <ivan@vald.es>