Migrate verify-govet-shadow make target to use the Go workspace
Update the function to use `go list $(workspace_relative_modules)` for
getting the list of packages.

Replace the array `skip_pkgs` with a Bash native associative array for
better readability, and simplicity of the code.

Remove the `govet_shadow_per_package` function, as it is not necessary
to iterate over the submodules anymore.

Signed-off-by: Ivan Valdes <ivan@vald.es>