Migrate scripts/fix.sh to use the Go workspace
Update the `mod_tidy_fix` function to iterate over modules and run `go
mod tidy` for each module directory. As it is not possible to execute
`go mod tidy` at the top-level of the repository.

Modify the `gofmt` call to leverage the Go workspace by removing the
iteration over submodules.

Signed-off-by: Ivan Valdes <ivan@vald.es>