Migrate cov make target to use the Go workspace
Update the `cov_pass` function in the `scripts/test.sh` file to remove
`run_for_module` and directly call `go_test`.

Remove the `modules` function from `scripts/test_lib.sh`, as it is not
needed anymore as the rest of the functions are now using the Go
workspace.

Update the implemenation of "not_test_pacakges" to not include the tools
submodules, as the original implementation doesn't include them.

Signed-off-by: Ivan Valdes <ivan@vald.es>