scripts/release: Use `go mod` instead of `go list`
Replace `go list -m` with `go mod edit -json`, as the latter can return
the same information. This will be helpful when the project migrates to
using a Go workspace, as it will return the current module defined in
go.mod rather than all the modules from the current directory (using a
workspace, the top-level go.mod will return all the child modules from
the repository).

Signed-off-by: Ivan Valdes <ivan@vald.es>