Add Kubernetes API coverage test consuming traces
It pulls a file from testdata with traces collected by Jaeger and then
iterates over them in order to group them into those which use the
contract interface, are known (and usually valid) uses outside of it.
Later it can further analyze the parameters used over the run of
Kubernetes conformance tests.

It requires patched Kubernetes
with:
https://github.com/AwesomePatrol/kubernetes/commit/5a18a66c06bad4350fb1fd52b3ca50ab885dbbf1
or at:
https://github.com/AwesomePatrol/kubernetes/tree/add-kubernetes-etcd-contract-tracker