mvcc: switch to fail point in test
This test is hard to provoke without inserting a sleep at the right place...
fortunately this is what failpoints can do for us.

Note that to my understanding, the user still needs to run `gofail enable` etc
in order to run this unit test with this delay inserted (i.e., I don't see that
we are using failpoints already as part of the unit testing flow for anything in
this package).