etcdserver: add size of the watchevent metrics
To improve observability of etcd v3. we counted the size of
WatchResponse sent by watchServer. Metrics are counted using
histograms, which can record the number of WatchResponses
sent and the size of WatchResponses. When there are a large
number of Watchers (especially when watching a lot of kvs),
sending WatchResponses will incur memory overhead. add this metric,
we can give users more information to find the root cause.

Signed-off-by: wangxiang <scottwangsxll@gmail.com>