[release-3.5]: Bump Go to 1.24.6
Signed-off-by: Milap Anwani <milapanwani@gmail.com>

tests: skip TestCorruptedBackupFileCheck test on big endian platforms

This skips the TestCorruptedBackupFileCheck test that uses a little endian
format file as test input and causes the test to fail for the wrong reason.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>

tests: skip TestCorruptedBackupFileCheck test on big endian platforms

This skips the TestCorruptedBackupFileCheck test that uses a little endian
format file as test input and causes the test to fail for the wrong reason.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>

[release-3.5]: Bump Go to 1.24.6

Signed-off-by: Milap Anwani <milapanwani@gmail.com>

[release-3.5]: Address govet 1.24 errors

Signed-off-by: Milap Anwani <milapanwani@gmail.com>

[release-3.5]: Address failing unit tests

Signed-off-by: Milap Anwani <milapanwani@gmail.com>

tests: skip TestCorruptedBackupFileCheck test on big endian platforms

This skips the TestCorruptedBackupFileCheck test that uses a little endian
format file as test input and causes the test to fail for the wrong reason.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>
Signed-off-by: Milap Anwani <milapanwani@gmail.com>

Update api/go.mod

Co-authored-by: Benjamin Wang <benjamin.wang@broadcom.com>
Signed-off-by: Milap Anwani <milapanwani@gmail.com>

[chore]: Update go.mod files as per suggestion

Signed-off-by: Milap Anwani <milapanwani@gmail.com>

[chore]: Update Errorf occurrences

Signed-off-by: Milap Anwani <milapanwani@gmail.com>

tests: skip TestCorruptedBackupFileCheck test on big endian platforms

This skips the TestCorruptedBackupFileCheck test that uses a little endian
format file as test input and causes the test to fail for the wrong reason.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>

tests: skip TestCorruptedBackupFileCheck test on big endian platforms

This skips the TestCorruptedBackupFileCheck test that uses a little endian
format file as test input and causes the test to fail for the wrong reason.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>