[release-3.4]: Bump Go to 1.24.6
Signed-off-by: Milap Anwani <milapanwani@gmail.com>

[release-3.4]: Address govet 1.24 errors

Signed-off-by: Milap Anwani <milapanwani@gmail.com>

tests: skip TestCorruptedBackupFileCheck test on big endian platforms

This skips the TestCorruptedBackupFileCheck test that uses a little endian
format file as test input and causes the test to fail for the wrong reason.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>

[chore]: Make changes as per suggestion

Signed-off-by: Milap Anwani <milapanwani@gmail.com>

tests: skip TestCorruptedBackupFileCheck test on big endian platforms

This skips the TestCorruptedBackupFileCheck test that uses a little endian
format file as test input and causes the test to fail for the wrong reason.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>

tests: skip TestCorruptedBackupFileCheck test on big endian platforms

This skips the TestCorruptedBackupFileCheck test that uses a little endian
format file as test input and causes the test to fail for the wrong reason.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>

[chore]: Make changes as per suggestion

Signed-off-by: Milap Anwani <milapanwani@gmail.com>