server/embed: Log EOF on DEBUG in TLS handshake
When EOF is encountered during TLS handshake, it is still logged as a
warning. This seems excessive and not really useful.

This patch ensures EOF is logged on debug only.

Unit tests are also added, but since the handler doesn't use an
interface for passing the TLS connection in, mocking that is hard. It's
achieved by using a private global variable, which is really not ideal,
but given the situation, it's the best tradeoff between the code being
pretty and complex.

Signed-off-by: Ondra Kupka <okupka@redhat.com>