[Robustness test] Create watches to cover more cases around the working window
There is a moving window for the revisions that are currently active. To be
more specific, there are revisions which have been compacted, revisions
that is available now, and revisions in the future.

During the robustness test, we would create watches to cover all 3 cases
at random moments during testing, to see that
- the watch created for compacted revisions should always fail
- the watch created on active or future revisions should always
  succeeded, and never breaks guarantees.

Signed-off-by: Chun-Hung Tseng <henrytseng@google.com>