Use depguard linter to forbid dependencies
Introduce the depguard linter to forbid required dependencies in the
submodules. This is a clean way to achieve what we had before in the
go.mod by using a replacement and pointing the forbidden package to a
virtual non-existent directory "FORBIDDEN_DEPENDENCY". Now, if a package
requires a forbidden dependency, the linter will fail, rather than
having Go fail to compile/fetch dependencies.

Signed-off-by: Ivan Valdes <ivan@vald.es>