Replace gomodguard standalone check with golanci-lint
Based on the discussion from #20721, and from commit
2edb08642c22c94a71a055d47ccf5f60d65e3490. The forbidden dependency
approach in the `go.mod` files was to avoid importing etcd v1.
Therefore, if we want to prohibit this dependency across all files, we
can keep it centralized in the golangci-lint configuration. Given that
the v2 client is also deprecated, we can add it to the forbidden
dependencies.

Follow up on the implementation from #20748.

Signed-off-by: Ivan Valdes <ivan@vald.es>