Silence BOM first run expected error
The tool github.com/appscodelabs/license-bill-of-materials fails during
the first run in a recently cloned/fresh repository. Because it captures
the output, and the first run shows the download of dependencies. In
8b0241630167ba48133c7b9a9bfb149e8af15c2e, we intentionally added a new
execution before evaluating whether the BOM is up to date. However, this
is polluting the logs as it has the expected failure. Redirecting the
output to /dev/null alleviates the issue.

Signed-off-by: Ivan Valdes <ivan@vald.es>