tests: Change max retries when removing a member from a cluster
In CI, the TestGateway and TestMixVersionsSnapshotByAddingMember
are flaky due to the TestMixVersionsSnapshotByAddingMember test
sometimes not closing the second etcd process. This happens if
the second process has not had enough time to become healthy according
to the logic in EtcdServer.mayRemoveMember.

Fix this by retrying member removal for twice the etcdserver.HealthInterval
in EtcdProcessCluster.CloseProc.

Signed-off-by: Jonathan Albrecht <jonathan.albrecht@ibm.com>