Run update-go-workspace late in the fix Makefile target
The file ./scripts/fix.sh is the one that updates the go modules to make
them tidy. However, we're running the update to the workspace before
running go mod tidy. Therefore, it is required to run make fix twice
right now. Therefore, because we can't change the order of prerequisites
(as scripts/fix.sh is not a prereq), call make update-go-workspace after
the fix.sh script.

Signed-off-by: Ivan Valdes <ivan@vald.es>