Replace force word splitting with a Bash array
Currently, we need to force word splitting, as we need to pass multiple
arguments, not a single one, to the command that will run using the
workspace modules.

To properly remove this exception, we need to use an array. The problem
with Bash arrays is that you can't return them in a function. So, the
way to build an array is by reference. Otherwise, we'll end up
rebuilding the arrays every time we need to call these functions.

Signed-off-by: Ivan Valdes <ivan@vald.es>